Rename eslint-plugin to eslint to remove ambiguity about being a console plugin
