+++ packages/common/.prettierrc.cjs
@@ -1,5 +1,5 @@
 /* eslint-env node */
 
 module.exports = {
-  ...require('@kubev2v/eslint-plugin/prettierrc.js'),
+  ...require('@kubev2v/eslint/prettierrc.js'),
 };
