+++ packages/eslint/.eslintrc.cjs
@@ -2,5 +2,5 @@
 
 module.exports = {
   root: true,
-  extends: ['plugin:@kubev2v/eslint-plugin/base'],
+  extends: ['plugin:@kubev2v/eslint/base'],
 };
