+++ packages/eslint/README.md
@@ -1,3 +1,3 @@
-# @kubev2v/eslint-plugin
+# @kubev2v/eslint
 
 Collection of linting configurations shared by `@kubev2v` packages.
