+++ packages/types/.eslintrc.cjs
@@ -2,7 +2,7 @@
 
 module.exports = {
   root: true,
-  extends: ['plugin:@kubev2v/eslint-plugin/typescript'],
+  extends: ['plugin:@kubev2v/eslint/typescript'],
   rules: {
     '@typescript-eslint/no-explicit-any': 'off',
     '@typescript-eslint/prefer-as-const': 'off',
