+++ packages/webpack/.eslintrc.cjs
@@ -2,5 +2,5 @@
 
 module.exports = {
   root: true,
-  extends: ['plugin:@kubev2v/eslint-plugin/typescript'],
+  extends: ['plugin:@kubev2v/eslint/typescript'],
 };
